+++ NEWS.md
@@ -1,5 +1,14 @@
 # NEWS for **spacesXYZ** package
 
+### Version 1.0-2  [2018-11-26]
+
+* added 4 functions for working with CCT, and a vignette for them
+* add function `daylightLocus()`
+* added function `uvfromxy()`
+* added function `standardxy()`
+* in `uvfromXYZ()`, changed name of argument `version` to `space`
+* in `DeltaE()` added CIE metrics for 1994 and 2000
+
 ### Version 1.0-1  [2018-07-06]
 
-initial version on CRAN
+* initial version on CRAN